requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# dobo_party/dobo_party/asgi.py: 12
# dobo_party/dobo_party/urls.py: 16,17
# dobo_party/dobo_party/views.py: 1,2
# dobo_party/dobo_party/wsgi.py: 12
# dobo_party/manage.py: 11
# dobo_party/meme_generator/admin.py: 1
# dobo_party/meme_generator/apps.py: 1
# dobo_party/meme_generator/models.py: 1
# dobo_party/meme_generator/tests.py: 1
# dobo_party/meme_generator/urls.py: 16,17
# dobo_party/meme_generator/views.py: 1,2
# dobo_party/urls.py: 2
Django == 3.2.9

# dobo_party/dobo_party/views.py: 3
# dobo_party/meme_generator/meme_generator.py: 1
# dobo_party/meme_generator/views.py: 3
Pillow == 7.0.0

# dobo_party/dobo_party/views.py: 7
python_decouple == 3.5

# dobo_party/dobo_party/views.py: 6
# dobo_party/meme_generator/meme_generator.py: 2
requests == 2.22.0
